更新 LauncherRightPanelViewController.m

+++ Natives/LauncherRightPanelViewController.m
@@ -10,7 +10,7 @@
 
 // 添加 C 函数声明 - 这些函数在 LauncherPreferences.m 或其他地方定义
 extern void setPrefString(NSString *key, NSString *value);
-extern void setPrefInt(NSString *key, int value);
+extern void setPrefInt(NSString *key, NSInteger value);
 
 @interface LauncherRightPanelViewController () <UIDocumentPickerDelegate>
